Remove handlers from Web. Add Path to properties.
The 'Web' module no longer adds event handlers to the div
representation of expressions.
It instead adds the 'Path' that connects a div with a subexrpession
to the divs property.
Adding the .clickable class, handling events and reading the
'Path' from the properties must be done by at a different place.